backend/requirements.txt
fastapi==0.110.1
starlette==0.37.2
uvicorn[standard]==0.25.0
pydantic==2.12.5
pydantic-settings==2.7.1
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
redis==5.2.1
fakeredis==2.27.0
httpx==0.28.1
python-dotenv==1.0.1
bcrypt==4.2.1
PyJWT==2.10.1
cryptography==44.0.0
resend==2.5.1
email-validator==2.2.0
pytest==8.3.4
